Made the light themes into their own component
